Alternative to #13
Still looks pretty bleak:

```
benchmark                               old MB/s     new MB/s     speedup
BenchmarkBlock8-32                      2509.41      2523.00      1.01x
BenchmarkAvx2/32KB-32                   2199.15      1765.44      0.80x
BenchmarkAvx2/64KB-32                   2936.65      2344.89      0.80x
BenchmarkAvx2/128KB-32                  3338.90      2774.66      0.83x
BenchmarkAvx2/256KB-32                  3558.00      3022.57      0.85x
BenchmarkAvx2/512KB-32                  3513.83      3104.90      0.88x
BenchmarkAvx2/1MB-32                    3433.49      3079.63      0.90x
BenchmarkAvx2/2MB-32                    3416.81      3061.47      0.90x
BenchmarkAvx2/4MB-32                    3425.56      3099.28      0.90x
BenchmarkAvx2/8MB-32                    3415.68      3102.60      0.91x
BenchmarkAvx2Parallel/32KB-32           31816.47     8515.98      0.27x
BenchmarkAvx2Parallel/64KB-32           38000.07     9303.00      0.24x
BenchmarkAvx2Parallel/128KB-32          41164.65     11504.55     0.28x
BenchmarkAvx2Parallel/256KB-32          43592.91     15237.84     0.35x
BenchmarkAvx2Parallel/512KB-32          44030.27     15926.92     0.36x
BenchmarkAvx2Parallel/1MB-32            44192.77     16602.93     0.38x
BenchmarkAvx2Parallel/2MB-32            44830.55     16490.63     0.37x
BenchmarkAvx2Parallel/4MB-32            44405.23     15724.88     0.35x
BenchmarkAvx2Parallel/8MB-32            43470.73     15546.07     0.36x
```
